move repair logic to after archives launched
